Refactor drizzle modules
