+++ cloudnetpy/products/drizzle.py
@@ -1,19 +1,13 @@
 """Module for creating Cloudnet drizzle product.
 """
-import logging
-import os
-from bisect import bisect_left
 from typing import Optional
 import numpy as np
 from numpy import ma
 from scipy.special import gamma
-import netCDF4
 from cloudnetpy import utils, output
-from cloudnetpy.categorize import DataSource
 from cloudnetpy.metadata import MetaData
-from cloudnetpy.products import product_tools as p_tools
-from cloudnetpy.products.product_tools import ProductClassification
 from cloudnetpy.products.drizzle_error import get_drizzle_error
+from cloudnetpy.products.drizzle_tools import DrizzleClassification, DrizzleSource, SpectralWidth, DrizzleSolver
 
 
 def generate_drizzle(categorize_file: str,
@@ -25,9 +19,9 @@ def generate_drizzle(categorize_file: str,
     cloud radar and lidar measurements. The results are written in a netCDF file.
 
     Args:
-        categorize_file (str): Categorize file name.
-        output_file (str): Output file name.
-        uuid (str, optional): Set specific UUID for the file.
+        categorize_file: Categorize file name.
+        output_file: Output file name.
+        uuid: Set specific UUID for the file.
 
     Returns:
         str: UUID of the generated file.
@@ -45,13 +39,11 @@ def generate_drizzle(categorize_file: str,
     drizzle_source = DrizzleSource(categorize_file)
     drizzle_class = DrizzleClassification(categorize_file)
     spectral_width = SpectralWidth(categorize_file)
-    drizzle_solver = DrizzleSolver(drizzle_source, drizzle_class,
-                                   spectral_width)
+    drizzle_solver = DrizzleSolver(drizzle_source, drizzle_class, spectral_width)
     derived_products = DrizzleProducts(drizzle_source, drizzle_solver)
     errors = get_drizzle_error(drizzle_source, drizzle_solver)
     retrieval_status = RetrievalStatus(drizzle_class)
-    results = {**drizzle_solver.params, **derived_products.derived_products,
-               **errors}
+    results = {**drizzle_solver.params, **derived_products.derived_products, **errors}
     results = _screen_rain(results, drizzle_class)
     results['drizzle_retrieval_status'] = retrieval_status.retrieval_status
     _append_data(drizzle_source, results)
@@ -63,278 +55,19 @@ def generate_drizzle(categorize_file: str,
     return uuid
 
 
-class DrizzleSource(DataSource):
-    """Class holding the input data for drizzle calculations.
-
-    Args:
-        categorize_file (str): Categorize file name.
-
-    Attributes:
-        mie (dict): Mie look-up table data.
-        dheight (float): Median difference of height array.
-        z (ndarray): 2D radar echo (linear units).
-        beta (ndarray): 2D lidar backscatter.
-        v (ndarray): 2D doppler velocity.
-
-    """
-    def __init__(self, categorize_file):
-        super().__init__(categorize_file)
-        self.mie = self._read_mie_lut()
-        self.dheight = utils.mdiff(self.getvar('height'))
-        self.z = self._convert_z_units()
-        self.beta = self.getvar('beta')
-        self.v = self.getvar('v')
-
-    def _convert_z_units(self):
-        """Converts reflectivity factor to SI units."""
-        z = self.getvar('Z') - 180
-        z[z > 0.0] = 0.0
-        return utils.db2lin(z)
-
-    def _read_mie_lut(self):
-        """Reads mie scattering look-up table."""
-        mie_file = self._get_mie_file()
-        nc = netCDF4.Dataset(mie_file)
-        mie = nc.variables
-        lut = {'Do': mie['lu_medianD'][:],
-               'mu': mie['lu_u'][:],
-               'S': mie['lu_k'][:],
-               'lwf': mie['lu_LWF'][:],
-               'termv': mie['lu_termv'][:]}
-        band = self._get_wl_band()
-        lut.update({'width': mie[f"lu_width_{band}"][:],
-                    'ray': mie[f"lu_mie_ray_{band}"][:],
-                    'v': mie[f"lu_v_{band}"][:]})
-        nc.close()
-        return lut
-
-    @staticmethod
-    def _get_mie_file():
-        module_path = os.path.dirname(os.path.abspath(__file__))
-        return '/'.join((module_path, 'mie_lu_tables.nc'))
-
-    def _get_wl_band(self):
-        """Returns string corresponding the radar frequency."""
-        radar_frequency = float(self.getvar('radar_frequency'))
-        wl_band = utils.get_wl_band(radar_frequency)
-        return '35' if wl_band == 0 else '94'
-
-
-class DrizzleClassification(ProductClassification):
-    """Class storing the information about different drizzle types,
-    child of  :class:`ProductClassification`.
-
-    Args:
-        categorize_file (str): Categorize file name.
-
-    Attributes:
-        is_v_sigma (ndarray): 2D array denoting finite v_sigma.
-        warm_liquid (ndarray): 2D array denoting warm liquid.
-        drizzle (ndarray): 2D array denoting drizzle presence.
-        would_be_drizzle (ndarray): 2D array denoting possible drizzle pixels.
-        cold_rain (ndarray): 1D array denoting profiles with melting layer.
-
-    """
-    def __init__(self, categorize_file):
-        super().__init__(categorize_file)
-        self.is_v_sigma = self._find_v_sigma(categorize_file)
-        self.warm_liquid = self._find_warm_liquid()
-        self.drizzle = self._find_drizzle()
-        self.would_be_drizzle = self._find_would_be_drizzle()
-        self.cold_rain = self._find_cold_rain()
-
-    @staticmethod
-    def _find_v_sigma(cat_file):
-        v_sigma = p_tools.read_nc_fields(cat_file, 'v_sigma')
-        return np.isfinite(v_sigma)
-
-    def _find_warm_liquid(self):
-        return (self.category_bits['droplet']
-                & ~self.category_bits['cold'])
-
-    def _find_drizzle(self):
-        return (~utils.transpose(self.is_rain)
-                & self.category_bits['falling']
-                & ~self.category_bits['droplet']
-                & ~self.category_bits['cold']
-                & ~self.category_bits['melting']
-                & ~self.category_bits['insect']
-                & self.quality_bits['radar']
-                & self.quality_bits['lidar']
-                & ~self.quality_bits['clutter']
-                & ~self.quality_bits['molecular']
-                & ~self.quality_bits['attenuated']
-                & self.is_v_sigma)
-
-    def _find_would_be_drizzle(self):
-        return (~utils.transpose(self.is_rain)
-                & self.warm_liquid
-                & self.category_bits['falling']
-                & ~self.category_bits['melting']
-                & ~self.category_bits['insect']
-                & self.quality_bits['radar']
-                & ~self.quality_bits['clutter']
-                & ~self.quality_bits['molecular'])
-
-    def _find_cold_rain(self):
-        return np.any(self.category_bits['melting'], axis=1)
-
-
-class SpectralWidth:
-    """Calculates corrected spectral width.
-
-    Removes the effect of turbulence and horizontal wind that cause
-    spectral broadening of the Doppler velocity.
-
-    Args:
-        categorize_file (str): Categorize file name.
-
-    Attributes:
-        categorize_file (str): Categorize file name.
-        width_ht (ndarray): Spectral width containing the correction for
-            turbulence broadening.
-
-    """
-    def __init__(self, categorize_file):
-        self.cat_file = categorize_file
-        self.width_ht = self._calculate_spectral_width()
-
-    def _calculate_spectral_width(self):
-        v_sigma = p_tools.read_nc_fields(self.cat_file, 'v_sigma')
-        try:
-            width = p_tools.read_nc_fields(self.cat_file, 'width')
-        except KeyError:
-            width = 0
-            logging.warning(f'No spectral width, assuming width = {width}')
-        sigma_factor = self._calc_v_sigma_factor()
-        return width - sigma_factor * v_sigma
-
-    def _calc_v_sigma_factor(self):
-        beam_divergence = self._calc_beam_divergence()
-        wind = self._calc_horizontal_wind()
-        actual_wind = (wind + beam_divergence) ** (2/3)
-        scaled_wind = (30*wind + beam_divergence) ** (2/3)
-        return actual_wind / (scaled_wind - actual_wind)
-
-    def _calc_beam_divergence(self):
-        beam_width = 0.5
-        height = p_tools.read_nc_fields(self.cat_file, 'height')
-        return height * np.deg2rad(beam_width)
-
-    def _calc_horizontal_wind(self):
-        """Calculates magnitude of horizontal wind.
-
-        Returns:
-            ndarray: Horizontal wind (m s-1).
-
-        """
-        atmosphere = p_tools.interpolate_model(self.cat_file, ['uwind', 'vwind'])
-        u_wind = atmosphere['uwind']
-        v_wind = atmosphere['vwind']
-        return utils.l2norm(u_wind, v_wind)
-
-
-class DrizzleSolver:
-    """Estimates drizzle parameters.
-
-    Args:
-        drizzle_source (DrizzleSource): The :class:`DrizzleSource` instance.
-        drizzle_class (DrizzleClassification): The :class:`DrizzleClassification` instance.
-        spectral_width (SpectralWidth): The :class:`SpectralWidth` instance.
-
-    Attributes:
-        params (dict): Dictionary of retrieved drizzle parameters 'Do', 'mu',
-            'S', 'beta_corr'.
-
-    """
-    def __init__(self, drizzle_source, drizzle_class, spectral_width):
-        self._data = drizzle_source
-        self._drizzle_class = drizzle_class
-        self._width_ht = spectral_width.width_ht
-        self._width_lut = -self._data.mie['width'][:]
-        self.params, self._dia_init = self._init_variables()
-        self._beta_z_ratio = self._calc_beta_z_ratio()
-        self._solve_drizzle(self._dia_init)
-
-    def _init_variables(self):
-        shape = self._data.z.shape
-        res = {'Do': np.zeros(shape), 'mu': np.zeros(shape),
-               'S': np.zeros(shape), 'beta_corr': np.ones(shape)}
-        return res, np.zeros(shape)
-
-    def _calc_beta_z_ratio(self):
-        return 2 / np.pi * self._data.beta / self._data.z
-
-    def _find_lut_indices(self, ind, dia_init, n_dia, n_widths):
-        ind_dia = bisect_left(self._data.mie['Do'], dia_init[ind], hi=n_dia - 1)
-        ind_width = bisect_left(self._width_lut[:, ind_dia], -self._width_ht[ind],
-                                hi=n_widths-1)
-        return ind_width, ind_dia
-
-    def _update_result_tables(self, ind, dia, lut_ind):
-        self.params['Do'][ind] = dia
-        self.params['mu'][ind] = self._data.mie['mu'][lut_ind[0]]
-        self.params['S'][ind] = self._data.mie['S'][lut_ind]
-
-    @staticmethod
-    def _is_converged(ind, dia, dia_init):
-        threshold = 1e-3
-        return abs((dia - dia_init[ind]) / dia_init[ind]) < threshold
-
-    @staticmethod
-    def _calc_dia(beta_z_ratio, mu=0, ray=1, k=1):
-        """ Drizzle diameter calculation.
-
-        Args:
-            beta_z_ratio (ndarray): Beta to z ratio, multiplied by (2 / pi).
-            mu (ndarray, optional): Shape parameter for gamma calculations. Default is 0.
-            ray (ndarray, optional): Mie to Rayleigh ratio for z. Default is 1.
-            k (ndarray, optional): Alpha to beta ratio . Default is 1.
-
-        Returns:
-            ndarray: Drizzle diameter.
-
-        References:
-            https://journals.ametsoc.org/doi/pdf/10.1175/JAM-2181.1
-
-        """
-        const = ray * k * beta_z_ratio
-        return (gamma(3 + mu) / gamma(7 + mu) * (3.67 + mu) ** 4 / const) ** (1 / 4)
-
-    def _solve_drizzle(self, dia_init):
-        drizzle_ind = np.where(self._drizzle_class.drizzle == 1)
-        dia_init[drizzle_ind] = self._calc_dia(self._beta_z_ratio[drizzle_ind], k=18.8)
-        n_widths, n_dia = self._width_lut.shape[0], len(self._data.mie['Do'])
-        max_ite = 10
-        for ind in zip(*drizzle_ind):
-            for _ in range(max_ite):
-                lut_ind = self._find_lut_indices(ind, dia_init, n_dia, n_widths)
-                dia = self._calc_dia(self._beta_z_ratio[ind] * self.params['beta_corr'][ind],
-                                     self._data.mie['mu'][lut_ind[0]],
-                                     self._data.mie['ray'][lut_ind],
-                                     self._data.mie['S'][lut_ind])
-                self. _update_result_tables(ind, dia, lut_ind)
-                if self._is_converged(ind, dia, dia_init):
-                    break
-                self._dia_init[ind] = dia
-            beta_factor = np.exp(2 * self.params['S'][ind] * self._data.beta[ind] *
-                                 self._data.dheight)
-            self.params['beta_corr'][ind[0], (ind[-1]+1):] *= beta_factor
-
-
 class DrizzleProducts:
     """Calculates additional quantities from the drizzle properties.
 
     Args:
-        drizzle_source (DrizzleSource): The :class:`DrizzleSource` instance.
-        drizzle_solver (DrizzleSolver): The :class:`DrizzleSolver` instance.
+        drizzle_source: The :class:`DrizzleSource` instance.
+        drizzle_solver: The :class:`DrizzleSolver` instance.
 
     Attributes:
         derived_products (dict): Dictionary containing derived drizzle products:
             'drizzle_N', 'drizzle_lwc', 'drizzle_lwf', 'v_drizzle', 'v_air'.
 
     """
-    def __init__(self, drizzle_source, drizzle_solver):
+    def __init__(self, drizzle_source: DrizzleSource, drizzle_solver: DrizzleSolver):
         self._data = drizzle_source
         self._params = drizzle_solver.params
         self._ind_drizzle, self._ind_lut = self._find_indices()
@@ -395,14 +128,13 @@ class RetrievalStatus:
     """Estimates the status of drizzle retrievals.
 
     Args:
-        drizzle_class (DrizzleClassification): The :class:`DrizzleClassification` instance.
+        drizzle_class: The :class:`DrizzleClassification` instance.
 
     Attributes:
-        drizzle_class (DrizzleClassification): The :class:`DrizzleClassification` instance.
-        retrieval_status (ndarray): 2D array containing drizzle retrieval status
-            information.
+        drizzle_class: The :class:`DrizzleClassification` instance.
+        retrieval_status (ndarray): 2D array containing drizzle retrieval status information.
     """
-    def __init__(self, drizzle_class):
+    def __init__(self, drizzle_class: DrizzleClassification):
         self.drizzle_class = drizzle_class
         self.retrieval_status = None
         self._get_retrieval_status()
@@ -424,7 +156,7 @@ class RetrievalStatus:
         self.retrieval_status[in_warm_liquid == 1] = 4
 
 
-def _screen_rain(results, classification):
+def _screen_rain(results: dict, classification: DrizzleClassification):
     """Removes rainy profiles from drizzle variables.."""
     for key in results.keys():
         if not utils.isscalar(results[key]):
@@ -432,7 +164,7 @@ def _screen_rain(results, classification):
     return results
 
 
-def _append_data(drizzle_data, results):
+def _append_data(drizzle_data: DrizzleSource, results: dict):
     """Save retrieved fields to the drizzle_data object."""
     for key, value in results.items():
         if key != 'drizzle_retrieval_status':
